# config/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Paris Pricing Backend App
host: 'localhost:10010'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /crear_plan/leer_excel:
    x-swagger-router-controller: excel_reader
    post:
      summary: Uploads a file.
      operationId: excel_reader
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: upfile
          type: file
          description: The file to upload.
          required: true
          x-mimetype: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/SuccessResponse'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  ErrorResponse:
    type: object
    required:
      - message
    properties:
      message:
        type: string
      code:
        type: string
      errors:
        $ref: '#/definitions/PartialErrors'
  SuccessResponse:
    type: object
    required:
      - info
    properties:
      info:
        $ref: '#/definitions/InfoResponse'
      data:
        $ref: '#/definitions/ExcelJSONResponse'
      PartialErrors:
        $ref: '#/definitions/PartialErrors'
  ExcelJSONResponse:
    type: array
    items:
      $ref: '#/definitions/ExcelJSONObject'
  ExcelJSONObject:
    type: object
  InfoResponse:
    type: object
    required:
      - status
    properties:
      message:
        type: string
      status:
        type: string
  PartialErrors:
    type: array
    items:
      type: string